style: 메인 페이지 반응형으로 수정
