--- Frontend/src/pages/Thread/style.css
@@ -1,118 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Helvetica,
-      Arial, sans-serif;
-  }
-  body {
-    color: #262626;
-    width: 100%;
-    font-size: 14px;
-    line-height: 18px;
-  }
-  .app {
-    background-color: #fafafa;
-    width: 100%;
-    max-width: 100%;
-    min-height: 100vh;
-    text-align: center;
-  }
-  .app__header {
-    position: sticky;
-    top: 0;
-    z-index: 9999;
-    background-color: white;
-    height: 54px;
-    display: flex;
-    align-items: center;
-    object-fit: contain;
-    border-bottom: 1px solid lightgray;
-  }
-  .app__headerWrapper {
-    max-width: 975px;
-    margin: 0 auto;
-    width: 100%;
-    padding: 0 20px;
-    display: flex;
-    align-items: center;
-  }
-  .app__headerButtons {
-    display: flex;
-    column-gap: 15px;
-  }
-  button {
-    font-size: 14px;
-    line-height: 18px;
-    cursor: pointer;
-    border: 1px solid transparent;
-  }
-  .primary__button {
-    background-color: #0095f6;
-    border-radius: 4px;
-    color: #fff;
-    padding: 5px 9px;
-    font-weight: 600;
-    text-align: center;
-    display: flex;
-    margin-left: auto;
-    margin-top: 20px;
-    margin-right: auto;
-  }
-  .login__button {
-    background-color: orange;
-    border-radius: 4px;
-    color: #fff;
-    padding: 5px 9px;
-    font-weight: 600;
-    text-align: center;
-    display: flex;
-    margin-left: auto;
-    margin-top: 20px;
-    margin-right: auto;
-  }
-  .text__button {
-    color: #0095f6;
-    font-weight: 600;
-    line-height: 28px;
-    background-color: transparent;
-  }
-  .app__signUp {
-    display: flex;
-    flex-direction: column;
-    margin: 28px 0;
-    row-gap: 15px;
-  }
-  
-  .timeline {
-    margin: 0 auto;
-    max-width: 500px;
-    padding: 10px 0px;
-    background-color: #fbfbfb;
-  }
-  
-  .addButton{
-    margin: 0 0 0 80%;
-  }
-  
-  .headerAva{
-    margin-left: 5%;
-  }
-  
-  .avatar {
-    border-radius: 10px;
-    margin-top: 5px;
-    margin-right: 10px;
-    width: 54px;
-    height: 54px;
-    object-fit: contain;
-  }
-
-  .filter_icon{
-    background-color: #e3e3e3;
-    color: #0095f6;
-  }
-
-  .clickable :hover{
-    cursor: pointer;
-  }
