+++ Frontend/src/pages/index.js
@@ -1,4 +1,4 @@
 export * from "./FeedView";
 export * from "./Login";
 export * from "./Profile";
-export * from "./Thread";
+export * from "./Main";
